trying to work around rails autoloading issues
In development, after making changes which trigger Rails to reload
classes, if I hit a page after the app reloads, an ArgumentError gets
raised with the following message:

	A copy of NavHighlighter::ApplicationHelper has been removed
	from the module tree but is still active!

I'm not sure what's causing this, but I suspect it's an issue with Rails
autoloading. I'm moving NavHighlighter::NavHilighting into its own file
to reflect what Rails autoloading is expecting in the hopes that this
will fix the problem.